[refactor][update] creation of navigation wrapper
