+++ src/App.js
@@ -2,12 +2,15 @@ import React from "react";
 import { Router } from "./router";
 import "./index.css";
 import { AuthProvider, VideoProvider } from "./context";
+import { NavigationWrapper } from "./components";
 
 function App() {
   return (
     <AuthProvider>
       <VideoProvider>
-        <Router />
+        <NavigationWrapper>
+          <Router />
+        </NavigationWrapper>
       </VideoProvider>
     </AuthProvider>
   );
