+++ src/components/index.jsx
@@ -1,6 +1,7 @@
-import { Navbar } from "./navbar/Navbar";
-import { Topbar } from "./topbar/Topbar";
+import { Navbar } from "./navigationWrapper/Navbar";
+import { Topbar } from "./navigationWrapper/Topbar";
 import { HeroContainer } from "./heroContainer/HeroContainer";
 import { PreviewCard } from "./previewCard/PreviewCard";
+import { NavigationWrapper } from "./navigationWrapper/NavigationWrapper";
 
-export { Navbar, Topbar, HeroContainer, PreviewCard };
+export { Navbar, Topbar, HeroContainer, PreviewCard, NavigationWrapper };
