--- src/components/topbar/Topbar.module.css
@@ -1,21 +1,0 @@
-.topbar__logo img {
-  width: 100%;
-  height: 100%;
-}
-
-.topbar__logo {
-  width: 187.5px;
-  height: 56px;
-}
-
-@media screen and (min-width: 768px) {
-  .topbar__logo {
-    width: 187.5px;
-  }
-}
-
-@media screen and (min-width: 1024px) {
-  .topbar__logo {
-    display: none;
-  }
-}
